Edit main content & images template
Change made to the central section of the page,
including padding and the use of flex to properly
order the elements. Class name "title" now changed
to "info". Template for images now added, tried
to add border-radius but results in a bug where
the corners disappear, might have to look into
it later